Fixes issue #181.
Due to recent changes (commit d8524b8), there exist two entries for root.
One for host %  (root@'%')
One for host localhost. (root@localhost)

As a result the prior UPDATE statement to change any host "%" to "localhost" fails, as it conflicts with the already existing entry root@localhost.
To 'block remote access' for root - the intended behavior according to the comment - we simply now delete the second entry for any host root@'%'.

Unrelated, the "IDENTIFIED BY" had to be removed from the "GRANT ALL" because it is invalid syntax for the recently updated Mysql version.